Modifications for nps-based simulation

+++ subsystems/imu/imu_nps.c
@@ -38,13 +38,15 @@ void imu_impl_init(void) {
 void imu_periodic(void) {
 }
 
-void imu_feed_accel(int32_t x, int32_t y, int32_t z){
+void imu_feed_accel(double x, double y, double z){
+	//Yep, we are feeding doubles into int32_t, same as the nps sim code on pc...
 	VECT3_ASSIGN(imu.accel_unscaled, x, y, z);
 	  // set availability flags...
 	  imu_nps.accel_available = TRUE;
 }
 
-void imu_feed_gyro(int32_t x, int32_t y, int32_t z){
+void imu_feed_gyro(double x, double y, double z){
+	//Yep, we are feeding doubles into int32_t, same as the nps sim code on pc...
 	RATES_ASSIGN(imu.gyro_unscaled, x, y, z);
 
 	  // set availability flags...
